refactor: move asset proxy helper under pwa
