+++ frontend/src/lib/pwa/serviceWorkerAssetProxy.test.ts
