+++ frontend/src/lib/pwa/serviceWorkerAssetProxy.ts
